Finalize the HTML file syntax
The html files are updated and edited with more applicable syntax such as ID, Class, and type selectors.

It is important to add the correct syntax html file for future use as a structure in building an organized website.

Most especially in this repository, the next step after the  HTML is CSS.